pyproject.toml
[project]
name = "hilbertsfc"
version = "0.1.0"
description = "Ultra-fast 2D and 3D Hilbert space-filling curve encode/decode kernels (NumPy + Numba)."
readme = "README.md"
authors = [{ name = "Remco Leijenaar" }]
requires-python = ">=3.12"
dependencies = ["numba>=0.59.0", "numpy>=1.26.0"]
license = "MIT"
license-files = ["LICENSE"]

keywords = [
    "hilbert",
    "space-filling-curve",
    "sfc",
    "numba",
    "numpy",
    "indexing",
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]

[project.urls]
Homepage = "https://remcofl.github.io/HilbertSFC/"
Source = "https://github.com/remcofl/HilbertSFC"
Documentation = "https://remcofl.github.io/HilbertSFC/"
Issues = "https://github.com/remcofl/HilbertSFC/issues"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
bench = [
    "hilbert-bytes>=0.6.0",
    "hilbertcurve>=2.0.5",
    "numpy-hilbert-curve>=1.0.1",
]
dev = [
    { include-group = "test" },
    { include-group = "typecheck" },
    "prek>=0.3.1",
    "ruff>=0.15.0",
    "nox>=2025.11.12",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.0",
    "mkdocstrings[python]>=0.30.0",
    "pymdown-extensions>=10.14.0",
]
notebooks = [
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "rich>=14.3.2",
    "numpy>=2.3.5",
]
scripts = [
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pillow>=12.1.0",
    "rich>=14.3.2",
]
test = ["pytest>=9.0.2"]
typecheck = ["pyright>=1.1.408", "pyrefly>=0.51.0", "ty>=0.0.15"]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "N", "NPY", "UP"]
ignore = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.ty]
src.include = ["src", "tests", "scripts", "notebooks"]

[tool.pyright]
include = ["src", "tests", "scripts", "notebooks"]
pythonVersion = "3.12"
typeCheckingMode = "basic"

[tool.pyrefly]
project-includes = ["src", "tests", "scripts", "notebooks"]
python-version = "3.12.0"
infer-with-first-use = false

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
markers = [
    "slow: exhaustive tests that may take longer (deselect with '-m \"not slow\"')",
]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
